Edited to reflect previous changes to the README in the MSBot root
I updated the README in the MSBot root folder a week or so back to replace localhost with endpoint, and add the appropriate command line switches per service. With this commit, I'm reflecting those changes in the README that is in the "docs" folder. Not sure if this file should be removed altogether though since it is duplicate content.