Anonymous blocks were extracted for the following files:
# msbot-init.ts

The parameter for `connectService` was extracted as `const connectedService` of type `IEndpointService`.
Additionaly, the property `description` was removed as it was not being used for any of the services.

# msbot-connect-file.ts

The parameter for `FileService` constructor was extracted as `const connectedService` of type `IFileService`.
Additionaly, the property `type` was added as it was needed by the interface.

# msbot-connect-endpoint.ts

The parameter for `EndpointService` constructor was extracted as `const connectedService` of type `IEndpointService`.
Additionaly, the property `type` was added as it was needed by the interface.